#84 - adapt front for new back model (task details) #84 - unified structures to db (ints instead of long where needed) #84 - added dummy results for one tournament #84 - added id generator for test data #84 - moved tournament details front files to more specific folder

+++ e-arbiter-tournament-results/src/main/java/pl/cyganki/results/model/database/Result.java
@@ -30,9 +30,9 @@ public abstract class Result {
     @Column(name = "earned_points")
     private String earnedPoints;
 
-    public long getEarnedPoints() {
+    public int getEarnedPoints() {
         return Arrays.stream(this.earnedPoints.split(","))
-                .mapToLong(Long::parseLong)
+                .mapToInt(Integer::parseInt)
                 .sum();
     }
